requirements.txt
PyQt6>=6.4.0
PyOpenGL>=3.1.6
PyOpenGL_accelerate>=3.1.6
tifffile>=2023.1.1
numpy>=1.24.0
pytest>=7.4.0
imageio>=2.28.0
imageio-ffmpeg>=0.4.8
